requirements.txt
fastapi
uvicorn
torch
transformers
diffusers
accelerate
imageio[ffmpeg]
numpy
Pillow
python-multipart
requests
sentencepiece
protobuf
